requirements.txt
amqp==2.6.1
asgiref==3.3.1
attrs==20.2.0
billiard==3.5.0.5
cattrs==1.0.0
celery==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1.7
django-allauth==0.44.0
django-block-snippets==2.0.1
django-chamber==0.6.1
django-core==1.4.1
django-crispy-forms==1.11.1
django-filter==2.4.0
django-haystack==0.0.0
django-is-core==2.21.23
django-pyston==2.13.1
django-rest-authemail==2.0.6
django-restframework==0.0.1
django-tastypie==0.14.3
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
facebook-sdk==3.1.0
filemagic==1.6
httpie==2.4.0
idna==2.10
import-string==0.1.0
importlib-metadata==3.7.0
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
kombu==4.3.0
Markdown==3.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.1.0
plotly==4.14.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pyparsing==2.4.7
PyPrind==2.11.2
PySocks==1.7.1
python-dateutil==2.8.1
python-facebook-api==0.7.1
python-mimeparse==1.6.0
python3-openid==3.2.0
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.11.0
retrying==1.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.4.0
